fix: glossary
